+++ legacy/legacy.txt
@@ -1,0 +1,1 @@
+This is the legacy folder, this is where old Files, which Towny no longer needs are kept for Reasons.
