--- src/ToDo.txt
@@ -1,6 +1,0 @@
-TODO:
-	Start using this list.	
-		
-	
-	
-
